+++ src/main/java/com/hotel/service/input/RoomInput.java
@@ -3,9 +3,11 @@ package com.hotel.service.input;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class RoomInput {
     private Integer roomNumber;
-    private Double roomPricePerDay;
+    private BigDecimal roomPricePerDay;
 }
